+++ src/app/app.component.ts
@@ -1,22 +1,14 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { HeaderComponent } from './header/header.component';
-import { FooterComponent } from "./footer/footer.component";
-import { NgFor, NgIf } from '@angular/common';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, HeaderComponent, FooterComponent, NgIf, NgFor],
+  imports: [RouterOutlet, HeaderComponent],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss'
 })
 export class AppComponent {
   title = 'mentoring-first-project';
-  showMain: boolean = true
-  newPages: number[] = [5, 4, 3, 2, 1]
-  
-  isShowMain(){
-    this.showMain = !this.showMain
-  }
 }
