--- src/app/footer/footer.component.html
@@ -1,54 +1,0 @@
-<footer class="footer">
-    <div class="footer-contacts">
-        <div class="footer-contacts-data">
-            <img class="footer-logo" src="/assets/logo-image.svg" alt="Logo">
-            <div class="footer-contacts-data-communication">
-                <div class="footer-contacts-data-communication-item">
-                    <img src="/assets/footer/phone-icon.svg" alt="Phone">
-                    <div class="data-communication-content">                    
-                        <h3 class="data-communication-content-title">Контактный телефон</h3>
-                        <p class="data-communication-content-text">+ 7 (812) 309-09-34</p>
-                        <p class="data-communication-content-text">+ 7 (965) 084-29-09</p>
-                    </div>
-                </div>
-                <div class="footer-contacts-data-communication-item">
-                    <img src="/assets/footer/email-icon.svg" alt="Email">
-                    <div class="data-communication-content">                    
-                        <h3 class="data-communication-content-title">Электронная почта</h3>
-                        <p class="data-communication-content-text">tustin78&#64;mail.ru</p>
-                        <p class="data-communication-content-text">info&#64;tustin.ru</p>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="line"></div>
-        <div class="footer-contacts-form">
-            <h3 class="footer-contacts-form-title">Остались вопросы?</h3>
-            <p class="footer-contacts-form-subtitle">Узнавайте первым о скидках, предложениях и событиях еженедельно в своем почтовом ящике. Отпишитесь в любое время одним щелчком мыши.</p>
-            <div class="footer-contacts-form-input">
-                <input class="form-input" type="text" placeholder="Введите свою почту">
-                <button type="submit" class="form-button">Отправить</button>
-            </div>
-        </div>
-    </div>
-    <div class="line-bottom"></div>
-    <div class="footer-content">
-        <div class="footer-content-navigation">
-            <nav class="footer-navigation">
-                <ul class="footer-navigation-list">
-                    <li class="footer-navigation-item">Главная</li>
-                    <li class="footer-navigation-item">О компании</li>
-                    <li class="footer-navigation-item">Каталог</li>
-                </ul>
-            </nav>
-            <div class="footer-content-social-network">
-                <img src="/assets/footer/facebook.svg" alt="Facebook">
-                <img src="/assets/footer/youtube.svg" alt="Youtube">
-                <img src="/assets/footer/whatsapp.svg" alt="Whatsapp">
-            </div>
-        </div>
-        <p class="footer-content-text">
-            © 2021-2022, EDE
-        </p>
-    </div>
-</footer>
