--- src/app/footer/footer.component.scss
@@ -1,166 +1,0 @@
-.footer{
-    padding: 30px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%;
-    overflow: hidden;
-    background: #4b565e;
-}
-
-.footer-contacts {
-    display: flex;
-    justify-content: space-between;
-    width: 1320px;
-}
-
-.footer-contacts-data {
-    display: flex;
-    flex-direction: column;
-    gap: 30px;
-}
-
-.footer-logo{
-    width: 60px;
-    height: 40px;
-}
-
-.footer-contacts-data-communication {
-    display: flex;
-    gap: 60px;
-}
-
-.footer-contacts-data-communication-item {
-    display: flex;
-    align-items: flex-start;
-    gap: 10px;
-}
-
-.data-communication-content {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-}
-
-.data-communication-content-title {
-    font-weight: 400;
-    font-size: 12px;
-    color: rgba(255, 255, 255, 0.6);
-}
-
-.data-communication-content-text {
-    font-weight: 500;
-    font-size: 16px;
-    color: #fff;
-}
-
-
-.line {
-    border: 1px solid #566169;
-    height: 230px;
-    width: 0;
-}
-
-.footer-contacts-form {
-    width: 530px;
-}
-
-.footer-contacts-form-title {
-    margin-bottom: 10px;
-    font-weight: 700;
-    font-size: 18px;
-    color: #fff;
-}
-
-.footer-contacts-form-subtitle {
-    margin-bottom: 30px;
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 171%;
-    color: rgba(255, 255, 255, 0.6);
-}
-
-.footer-contacts-form-input {
-    position: relative; 
-    display: flex; 
-    align-items: center;
-    width: 534px;
-}
-
-.form-input {
-    padding: 0px 6px 0px 24px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 534px;
-    height: 48px;
-
-    background: #4b565e;
-    border: 1px solid #f0ba4e;
-    border-radius: 30px;
-
-    padding-left: 40px; 
-    background-image: url('/assets/footer/email-icon.svg');
-    background-repeat: no-repeat;
-    background-position: 15px center; 
-    background-size: 16px 16px; 
-}
-
-.form-input::placeholder{
-    font-weight: 400;
-    font-size: 12px;
-    color: rgba(255, 255, 255, 0.6);
-}
-
-.form-button {
-    position: absolute;
-    right: 10px;
-    width: 94px;
-    height: 36px;
-    border-radius: 30px;
-
-    font-weight: 700;
-    font-size: 10px;
-    text-align: center;
-    background-color: #f0ba4e;
-    color: #fff;
-}
-
-.line-bottom{
-    border-bottom: 1px solid #566169;
-    width: 1966px;
-}   
-
-.footer-content {
-    padding-top: 35px;
-    display: flex;
-    justify-content: space-between;
-    align-items: end;
-    width: 1320px;
-}
-
-.footer-navigation {
-    margin-bottom: 30px;
-}
-
-.footer-navigation-list {
-    display: flex;
-    gap: 30px;
-}
-
-.footer-navigation-item {
-    font-weight: 400;
-    font-size: 14px;
-    color: #fff;
-}
-
-.footer-content-social-network {
-    display: flex;
-    gap: 20px;
-}
-
-.footer-content-text {
-    font-weight: 400;
-    font-size: 14px;
-    color: rgba(255, 255, 255, 0.6);
-}
