+++ src/app/interfaces/users.interface.ts
@@ -1,0 +1,11 @@
+export interface User {
+    id: number;
+    name: string;
+    username: string;
+    email: string;
+    phone: string;
+    website: string;
+    address: {
+      city: string;
+    }
+  }
