+++ src/app/todos-list/todo-card/todo-card.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { Todo } from '../todos-list.component';
+import { Todo } from '../../interfaces/todos.interface';
 
 @Component({
   selector: 'app-todo-card',
