+++ src/app/users-api.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
-import { User } from './user-list/user-list.component';
+import { User } from './interfaces/users.interface';
 
 @Injectable({
   providedIn: 'root'
